requirements.txt
Flask==0.12.2
Flask-Cors==3.0.6
Flask-RESTful
requests==2.19.1
pyasn1>=0.1.3
pyCrypto==2.6.1
leveldb==0.193
cryptography==1.8.1
grpcio==1.3.5
grpcio-tools==1.3.5
coloredlogs
verboselogs
GitPython>=2.1.1
fluent-logger>=0.5.0
protobuf>=3.2.0
gunicorn
yappi